fix(components/atom/input): useMask hook refactor
